[Doc][Format] Format help document support using Enter Key
Kudu uses Function:AppendHardWrapped to format a description
of a command, which will split the string when it's length is
more than 78. But this function don't handle Enter key in the
string, the formatted result is indent.

For example:
Original string:
"Attempts to create on-disk metadata that can be used by master.\n"
"It also can be used by a replicated tserver\n"

Old function output:
Attempts to create on-disk metadata that can be used by master.
It also can be
used by a replicated tserver

New function output:
Attempts to create on-disk metadata that can be used by master.
It also can be used by a replicated tserver

The new Function will split the string using '\n', then split
every line when it's length is more than 78.

Change-Id: I1bd15bd2de292f534e90ac24e925bf605ddf6d7b
Reviewed-on: http://gerrit.cloudera.org:8080/18566
Tested-by: Kudu Jenkins
Reviewed-by: Yifan Zhang <chinazhangyifan@163.com>