[tests] make the startup page tests more robust
I noticed that the startup page tests reliably fail on my test machine.
It turned out that there was KUDU_HOME variable set in the environment,
so the embedded webserver rendered the pages according to the Mustache
mappings found under $KUDU_HOME/www, while the test scenarios expect all
pages to be pre-rendered.

This patch updates corresponding test scenarios to reliably work even if
the KUDU_HOME environment is set and points to the directory with
the Mustache mapping files.

Change-Id: I2e56cd49eb2e00c5666e4214c7ca4246386e8b9e
Reviewed-on: http://gerrit.cloudera.org:8080/18606
Tested-by: Kudu Jenkins
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
Reviewed-by: Attila Bukor <abukor@apache.org>