KUDU-2671 update Add{ExclusiveUpper,Lower}BoundPartitionKeyRaw
Since we don't allow for variable number of hash dimensions in
range-specific hash schemas for a table, this patch updates the code
in KuduScanner::AddExclusiveUpperBoundPartitionKeyRaw() and
KuduScanner::AddLowerBoundPartitionKeyRaw() to enable using them
for tables having custom hash schemas per range.

This is a follow-up to 6998193e69eeda497f912d1d806470c95b591ad4.

Change-Id: I1323e494733c9c082dedc12d941811f4175dbeab
Reviewed-on: http://gerrit.cloudera.org:8080/18640
Reviewed-by: Mahesh Reddy <mreddy@cloudera.com>
Tested-by: Kudu Jenkins
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
Reviewed-by: Attila Bukor <abukor@apache.org>