KUDU-3385 Integrate Kudu with Ranger KMS
To make data at rest encryption actually secure, the cluster key needs
to be generated and stored securely. Hadoop/Ranger KMS was chosen for
this purpose, which are API compatible.

This patch implements a Ranger KMS key provider and the Ranger client,
and also wraps up initialization of keys and policies in MiniRanger and
MiniRangerKMS. KeyProvider interface and its default mplementation had
to be moved into fs/ from server/ to avoid introducing circular
dependencies.

It also removes the 'hidden' tag from the encryption-related flags as it
is now safe to use.

Change-Id: I681bed35f44cd03c1db69847c04faa460745f206
Reviewed-on: http://gerrit.cloudera.org:8080/18692
Reviewed-by: Alexey Serbin <alexey@apache.org>
Tested-by: Attila Bukor <abukor@apache.org>