KUDU-2671 show custom hash schema in formatRangePartition
This patch introduces a new method targeted for Impala for formatted
output of table's range partitions along with hash their schemas.
Hash schema information is appended to each range partition.

Several test scenarios in TestKuduTable are updated to provide coverage
for the functionality of the newly introduced method
KuduTable.getFormattedRangePartitionsWithHashSchema().

Change-Id: Ia1a014ad820b2d73ea09fe3f67baefe9d0acb7aa
Reviewed-on: http://gerrit.cloudera.org:8080/18693
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Kurt Deschler <kdeschle@cloudera.com>
Reviewed-by: Alexey Serbin <alexey@apache.org>