KUDU-2671 fix handling ranges with table-wide hash schemas
This patch fixes a bug/typo introduced recently: adding a range with
the table-wide hash schema would be a no-op when
--enable_per_range_hash_schemas was set 'true'.

I added new test scenarios to cover the regression and verified
that, as expected, they would fail without the fix.

This is a follow-up to f55dd22bfea4beee99d72891efbbc67307b19d1e.

Change-Id: I908e9654ed856920d483ce63a546946cbac0a641
Reviewed-on: http://gerrit.cloudera.org:8080/18696
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Mahesh Reddy <mreddy@cloudera.com>
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
Reviewed-by: Attila Bukor <abukor@apache.org>