KUDU-3371 [fs] make LogBlockManager a base class
This patch makes LogBlockManager a base class for the newly
added LogBlockManagerNativeMeta.
FileMetaLogBlockManager is the log-backed block manager which
manages the sequential written file to store containers' metadata,
it is how we do as before.

Most of the member functions are left in the base class LogBlockManager
to avoid too large updating on existing code, I'm planning to do
that in next patches.

Change-Id: I59c1287a9539668e0c08036452bb96a3006ed356
Reviewed-on: http://gerrit.cloudera.org:8080/18709
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <alexey@apache.org>