KUDU-2671  Make WebUI compatible with custom hash schema
This patch updates the /table?id=<table_id> page in the Kudu master
WebUI to show custom hash schemas in the sections of:

1. Partition Schema
The custom hash schema if present for a particular range is displayed
right beside the range schema. Different dimensions of the hash
schema are comma separated.

2. Detail
There are new columns to identify if a particular partition has
custom or table wide hash schema, display the hash schema and the hash
partition id of the partition.

The Kudu tablet server WebUI's pages /tablets and
/tablet?id=<tablet_id> are also tested to reflect the custom hash
schema or table wide hash schema accordingly.

Below are the screenshots of the WebUI after the changes
Master WebUI:
https://i.imgur.com/O4ra4JA.png
Tablet server WebUI:
https://i.imgur.com/BxdfsYt.png
https://i.imgur.com/l2wA08Q.png

Change-Id: Ic8b8d90f70c39f13b838e858c870e08dacbdfcd3
Reviewed-on: http://gerrit.cloudera.org:8080/18712
Reviewed-by: Alexey Serbin <alexey@apache.org>
Tested-by: Kudu Jenkins