KUDU-2671 fix double-free in new scenario of scan_token-test
Changelist b746978c7 introduced a bug in the updated CountRowsSeq()
function -- the mistake was calling delete twice on KuduScanToken
pointers.  That lead to reading garbage data, so the newly introduced
test scenario ScanTokensWithCustomHashSchemasPerRange was failing if
enabled.  I didn't pay enough attention to the updated test code when
reviewing it since the new test scenario was disabled in the original
patch b746978c7.  This patch fixes the mistake.

Change-Id: I227f1dadf9d5b4d3b209570716cde7bda74c6b25
Reviewed-on: http://gerrit.cloudera.org:8080/18714
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Khazar Mammadli <mammadli.khazar@cloudera.com>
Reviewed-by: Mahesh Reddy <mreddy@cloudera.com>
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
Reviewed-by: Attila Bukor <abukor@apache.org>