KUDU-2671 fix consistency check in KuduScanner::NextBatch()
KuduScanner::Data::OpenTablet() can return earlier when the metacache
returns a tablet covering the partition key range past the partition key
provided.  In such case, the 'proxy_' member of KuduScanner::Data isn't
set, and that's expected since the upper level code should continue with
the next tablet.  The new test sub-scenarios introduced in [1] triggered
the consistency check on the 'data_->proxy_' member in
KuduScanner::NextBatch(), so this patch addresses the issue.

Also, this patch re-enables a few test sub-scenarios introduced in
[1] and [2].

[1] https://github.com/apache/kudu/commit/b746978c71ce4a95b69d49c43d0ac852909a8b4e
[2] https://github.com/apache/kudu/commit/1889d4c44385fec5efeeb2d287d9ab7a3544dcfe

Change-Id: Icfbdfac46f35d4d143f37845d158795bf1793da7
Reviewed-on: http://gerrit.cloudera.org:8080/18715
Reviewed-by: Khazar Mammadli <mammadli.khazar@cloudera.com>
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Mahesh Reddy <mreddy@cloudera.com>
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
Reviewed-by: Attila Bukor <abukor@apache.org>