KUDU-3379 Add column comments to table describe
This patch adds -show_column_comment flag to the 'kudu table describe'
CLI tool. If there are column comments specified, this can be used to
show them.

Schema and ColumnSchema use the ToStringMode enum to control how they
are stringified. Previously these were used to signal a single mode of
the stringification. Now both are populated through bit shifting. This
way the enum can be used as a holder of multiple binary flags. We can
check for the presence of a flag through 'bitwise and' operator, and set
a given flag through the 'bitwise or' operator.

A new test case is added to check whether the column describing flags
work, alone and together as well.

Change-Id: Id8b751bd821a5f50490caa6f1aaf1fc767de0222
Reviewed-on: http://gerrit.cloudera.org:8080/18903
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <alexey@apache.org>