[Client] Add query id to trace the whole query process
In Impala, an SQL is translated into a query plan. The
query plan contains multiple execution units and is
executed at different hosts of a cluster in a distributed
manner. There exists an unique query id to trace the
execution process of every execution unit. It is useful
in analysis of stuck, failed, and otherwise non-performant
queries. Query id can associate all log records in different
hosts, which belongs to one query.

Scanning a Kudu table is split into multiple scan operations
on appropriate tablets, one scanner per tablet. When scanning
a Kudu table using Impala, corresponding scan nodes appear
for every scanner in the Impala's query plan.  Scanner id is
used to trace the scan operation on a tablet.

But there is a gap between Impala and Kudu. We can't trace the
whole scanning process in Kudu of a query originated from Impala.
There is not unique id to associate the execution of all scanners
in Kudu.

Therefore, with this patch it is now possible to post Impala's
query identifier to Kudu, so it can be associated with Kudu's
scanner id. Then we can trace a query from Impala to Kudu for
troubleshooting and debugging.

Change-Id: I9dbae801596726fec1c85ee547128da3179345d9
Reviewed-on: http://gerrit.cloudera.org:8080/18846
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Alexey Serbin <alexey@apache.org>