[threadpool] Fix unsafe behaviour when SchedulerThread shutdown
When shutting down a thread pool with SchedulerThread, thread pool's variable
'scheduler_' is deleted and set to nullptr, at the same time, if thread pool
token is not shutdown and a task is scheduled on the token, this may cause
a coredump because 'scheduler_' is a nullptr.

This patch fixes the bug, adding a mutex lock in class 'ThreadPool' to
protect variable 'scheduler_'.

Change-Id: I021422f7e51e1007c5bdbc877ab445f70ba12357
Reviewed-on: http://gerrit.cloudera.org:8080/18867
Tested-by: Kudu Jenkins
Reviewed-by: Yingchun Lai <acelyc1112009@gmail.com>