[tools] fix backward compatibility in `kudu table delete`
[1] introduced incompatibility into the `kudu table delete` CLI tool
by setting --reserve_seconds to non-zero by default.  With that,
the tool started trying to soft-delete tables by default, which lead
to unexpected failures when HMS integration was enabled.

This patch addresses the issue by changing the default value of
--reserve_seconds to 0.  In addition, this patch adds a few test
scenarios to check that dropping a table works as expected when HMS
integration is enabled after the soft-delete functionality for tables
has been introduced.

This is a follow-up to [1].

[1] https://github.com/apache/kudu/commit/7b6b6b636818d3e22a3939fde77689dce84e88b2

Change-Id: I360d81c2f64ac07cd342ddb882aa72ee50b820bc
Reviewed-on: http://gerrit.cloudera.org:8080/18932
Tested-by: Kudu Jenkins
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
Reviewed-by: Attila Bukor <abukor@apache.org>