[tools] mark long-running scenarios in kudu-tool-test
I noticed that it takes very long time for kudu-tool-test to complete.
It turned out the test suite contains many scenarios that run with
HMS integration enabled and it takes a long time to start up the
HMS process.

This patch marks the scenarios that tend to run over 3 seconds
as slow, so they are skipped unless the KUDU_ALLOW_SLOW_TESTS
environment variable is set to true/yes/1.

The run time of the kudu-tool-test with this patch looks like the
following on a contemporary Linux machine:

Without KUDU_ALLOW_SLOW_TESTS:
  [==========] 128 tests from 8 test suites ran. (86355 ms total)
  [  PASSED  ] 88 tests.
  [  SKIPPED ] 40 tests, listed below:
  ...

With KUDU_ALLOW_SLOW_TESTS=1:
  [==========] 128 tests from 8 test suites ran. (842684 ms total)
  [  PASSED  ] 128 tests.

Change-Id: I1893741fada185d819128396c88f4ac99f43ed40
Reviewed-on: http://gerrit.cloudera.org:8080/18947
Reviewed-by: Mahesh Reddy <mreddy@cloudera.com>
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Yifan Zhang <chinazhangyifan@163.com>