[scripts] add HTTP address option for start_kudu.sh
This patch adds a new option to customize the IP address of the Kudu's
embedded HTTP server.  A couple of usage examples:

   ../../src/kudu/scripts/start_kudu.sh --webhost 192.168.1.37
   ../../src/kudu/scripts/start_kudu.sh -W 192.168.1.37

I also took the liberty to update the description of the flags to add
default values for each and update the naming of the flags that specify
the port numbers for the RPC and the HTTP endpoints while keeping the
script backwards-compatible
(i.e. it still understands the prior naming of the corresponding flags):

  * rpc-master  --> rpc-port-master
  * rpc-tserver --> rpc-port-tserver

Change-Id: I468499c5df56e958f926053d485b7d0a08c63905
Reviewed-on: http://gerrit.cloudera.org:8080/18948
Tested-by: Kudu Jenkins
Reviewed-by: Yifan Zhang <chinazhangyifan@163.com>