[cfile] style-guide updates on cfile_reader.{h,cc}
I was looking into the code of the CFileReader::ReadBlock() method while
investigating one compaction issue and took the liberty to update the
code in the corresponding files, making it compliant with the project's
style guide.  A follow-up patch makes the code in cfile_reader.{h,cc}
more robust.

There are no functional changes in this patch.

Change-Id: Ib0e900271e7127ea8bbc2c949761f5339740c121
Reviewed-on: http://gerrit.cloudera.org:8080/18964
Reviewed-by: Yingchun Lai <acelyc1112009@gmail.com>
Tested-by: Alexey Serbin <alexey@apache.org>