[compaction] style guide-related updates
I was debugging one compaction issue and found that compaction-related
code in compaction-test.cc isn't robust enough.  This patch just updates
the code in compaction.{h,cc} and compaction-test.cc to conform with the
project's style guidelines.  A follow-up patch is to make the code
in compaction-test.cc more robust.

There are no functional changes in this patch.

Change-Id: I4ef1981dad33a6bc2d0911630fd73c95b91e45fe
Reviewed-on: http://gerrit.cloudera.org:8080/18974
Tested-by: Kudu Jenkins
Reviewed-by: Yifan Zhang <chinazhangyifan@163.com>