KUDU-3403 Enable kudu CLI to accept specific leader to step down
Accept UUID as an argument to identify the leader
and deduce other information like host, port, etc.
This can be useful in scenario when master is out of sync and
user wants to choose a specific leader to step down instead of
relying on information about leader in the tablet configuration
known to master.

Change-Id: I40569faa40a8173c51504c7567aa84a6ae1fb64a
Reviewed-on: http://gerrit.cloudera.org:8080/19024
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <alexey@apache.org>