KUDU-3416: Fix javadoc build failure on non-Unicode systems
For a build of Kudu inside a Docker image where Unicode
is not the default, the :kudu-client:javadoc task fails with:

AsyncKuduClient.java:926: error: unmappable character for encoding ASCII

This is complaining about some Unicode parentheses in
AsyncKuduClient.java. To fix this issue, this changes the
Unicode parentheses to regular ASCII parentheses.

Testing:
 - Tested by hand on a non-Unicode system

Change-Id: Ia2c2b2e20e6b6b0d649bedf5f2a0341bb62f2b19
Reviewed-on: http://gerrit.cloudera.org:8080/19196
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <alexey@apache.org>