KUDU-1945 Auto-Incrementing Column
This patch adds a new column specification named auto_incrementing.
The ColumnSchema of this new column is INT64 and not UINT64 as
impala doesn't support UINT64. These columns are populated on the
server side with a monotonically increasing counter. This counter
is local to every tablet i.e. each tablet has a separate auto
incrementing counter. This is a step towards having tables with
non unique primary keys or in case of tables with just one tablet,
a table wide unique key.
Upon receiving a write request the leader replica:
1. Fills in the replicate message with the auto incrementing counter
   value for the first write op.
2. Populates the auto incrementing key into the rows being inserted
   during the prepare phase.
3. Sends out the replicate message with the auto incrementing counter
   and the original write request. The followers perform the same
   set of steps to populate the auto incrementing column.
It also adds the c++ client side changes needed for basic writes
and reads of this column to perform end-to-end tests.

Change-Id: I1dbde9095da78f6d1bd00adcc0a6e7dd63082bbc
Reviewed-on: http://gerrit.cloudera.org:8080/19097
Reviewed-by: Wenzhe Zhou <wzhou@cloudera.com>
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Alexey Serbin <alexey@apache.org>